+++ src/lib/core/account.ts
@@ -1,11 +1,11 @@
 import { accountsStorage } from '$lib/core/data-storage';
 import AvatarManager from '$lib/core/managers/avatar';
 import XMPPManager from '$lib/core/managers/xmpp';
-import Legendary from '$lib/utils/legendary';
+import Legendary from '$lib/core/legendary';
 import { get } from 'svelte/store';
 import type { AccountData } from '$types/accounts';
 import DeviceAuthManager from '$lib/core/managers/device-auth';
-import AutoKickBase from '$lib/core/managers/autokick-base';
+import AutoKickBase from '$lib/core/managers/autokick/base';
 
 export default class Account {
   static changeActiveAccount(id: string | null) {
