+++ src/lib/core/managers/autokick/manager.ts
@@ -4,10 +4,10 @@ import { getResolvedResults, sleep } from '$lib/utils/util';
 import type { AccountData } from '$types/accounts';
 import MatchmakingManager from '$lib/core/managers/matchmaking';
 import { EpicEvents } from '$lib/constants/events';
-import AutoKickBase from '$lib/core/managers/autokick-base';
+import AutoKickBase from '$lib/core/managers/autokick/base';
 import PartyManager from '$lib/core/managers/party';
-import claimRewards from '$lib/utils/autokick/claim-rewards';
-import transferBuildingMaterials from '$lib/utils/autokick/transfer-building-materials';
+import claimRewards from '$lib/core/managers/autokick/claim-rewards';
+import transferBuildingMaterials from '$lib/core/managers/autokick/transfer-building-materials';
 import { accountsStorage, settingsStorage } from '$lib/core/data-storage';
 import type { PartyData } from '$types/game/party';
 import { get } from 'svelte/store';
